Make bump_index usize instead of u8
Now there won't be an overflow if a board ever has more than 255 threads.

+++ src/actors/fetcher/mod.rs
@@ -296,7 +296,7 @@ impl Handler<FetchThreads> for Fetcher {
                         acc
                     });
                     for (i, thread) in threads.iter_mut().enumerate() {
-                        thread.bump_index = i as u8;
+                        thread.bump_index = i;
                     }
                     if threads.is_empty() {
                         Err(FetchError::EmptyData)
